move confirm email screen to its own file
